🔐 Fix auth typing
